refactor: déplace la restoration de la dernière simulation  + envoi de stats dans une lib externe
